fix packaging and imports and introduce tests with pytest. still issues with celery worker.
