+++ .github/workflows/pytest.yml
@@ -24,5 +24,4 @@ jobs:
           if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
       - name: Test with pytest
         run: |
-          cd application
-          pytest
+          python -m pytest
